pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tiny-lab"
version = "5.0.0"
description = "Plan-driven AI research loop"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
mlx = ["mlx>=0.29", "numpy>=2.0"]
dev = ["pytest>=8.0"]

[project.scripts]
tiny-lab = "tiny_lab.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tiny_lab"]
